Se agregan scripts para impactar en las tablas RetailChannelProfile

+++ CreateLocalConfigFile.ps1
@@ -1,34 +1,55 @@
-function GetCertThumbprint {
+function GetWebSiteCertThumbprint {
     param (
         [string]$webSite = "AOSService"
     )
-    
     # Obtiene la información del sitio web
     $sitio = Get-WebSite -Name $webSite
 
     # Obtiene el binding principal
     $bindings = $sitio.bindings
     $binding = $bindings.Collection[0]
+    
     # Obtiene el certificado asociado al binding
     $certificado = $binding.certificateHash
-    # Imprime el thumbprint del certificado
+    
     return $certificado
+}
+function GetWebSiteUrl {
+    param (
+        [string]$webSite
+    )
+    # Obtiene la información del sitio web
+    $sitio = Get-WebSite -Name $webSite
+
+    # Obtiene el binding principal
+    $bindings = $sitio.bindings
+    $binding = $bindings.Collection[0]
+    
+    # Usar una expresión regular para extraer el host
+    $hostSegment = [regex]::Match($binding.bindingInformation, ':\d+:(.+)$').Groups[1].Value
+    $protocol = $binding.protocol
+    $url =  $protocol + "://" + $hostSegment
+    return $url
 }
 
 # Crear archivo
 $hostname = $env:COMPUTERNAME
 $jsonFile = ".\ConfigFiles\$hostname.json"
-Copy-Item ".\ConfigFiles\SAMPLE_Config_By_Env_(DuplicateAndRename).json" $jsonFile
+if (-not (Test-Path $jsonFile)) {
+    Copy-Item ".\ConfigFiles\SAMPLE_Config_By_Env_(DuplicateAndRename).json" $jsonFile
+}
 
 # Cargar archivo
 $json = Get-Content $jsonFile | ConvertFrom-Json
 
 # Seteo de configuraciones
-$urlEnvironment = Get-D365Url
-$urlEnvironment = $urlEnvironment.Url
-$json.RetailServerURL = "$urlEnvironment/RetailServer/Commerce"
-$json.Thumbprint = GetCertThumbprint("AOSService")
-$json.EnvironmentId
+$RetailServerURL = GetWebSiteUrl('RetailServer')
+$json.RetailServerURL = "$RetailServerURL/Commerce"
+$CPOSUrl = GetWebSiteUrl('RetailCloudPos')
+$json.CPOSUrl = $CPOSUrl
+$json.Thumbprint = GetWebSiteCertThumbprint("AOSService")
+# TODO Falta averiguar como obtener este dato
+# $json.EnvironmentId = 
 
 # Guardado del archivo
 $json | ConvertTo-Json | Out-File $jsonFile
